fix: fix file management
