--- src/features/profile/subscriptions/index.ts
@@ -1,1 +1,0 @@
-export * from './model/types/services'
